14.06 testing load

+++ WebRoutes/Dtos/RequestDtos/LocationCreateRequestDto.cs
@@ -6,9 +6,9 @@ public class LocationCreateRequestDto
     
     public string? LocationDescription { get; init; }
     
-    public required double Latitude { get; init; }
+    public required float Latitude { get; init; }
     
-    public required double Longitude { get; init; }
+    public required float Longitude { get; init; }
     
     public IFormFile? LocationImage { get; init; }
 }
